refactor(ol_openedx_events_handler): flatten tasks/ package into tasks.py
